Move tests in spec.rs to tests/
